Restores explicit http status block with optional body in output from ktor.
This had been modelled as a pseudo domain class (`Response`) that wrapped header and body along with an extension function to the `ApplicationCall`. However, that
approach conflated http concerns and the chosen approach of wrapping up the domain concerns in a sort of JSON "envelope" with "normal" code.

As these concerns are all about how the domain logic is mapping into specific outputs, we now go for a simple Ktor plugin that takes responsibility for:
* mapping `null` responses to 404
* wrapping simple status code responses so that they can be serialised to JSON without the `ContentNegotiation` plugin intervening
* wrapping normal responses into a `statusCode` and `body`
* turning `IllegalArgumentExceptions` into 400 with an additional error message

`ApplicationTest` is extended to validate these cases.