Updates kotlin to 1.9.0, ktor to 2.3.2, adopts simple plugin to handle json status code and body wrapping (#5)
* updates dependencies other than ktor

* updates kotlin to 1.9.0, updates ktor to last 1.x version, fixes resulting compile errors in MeterReadingControllerTest

* changes last junit anotation to kotlin test library one

* upgrades ktor to 2.3.2, does not yet adopt tests

* Adjusts the application tests to use the new ktor v2 framework and simplifies the controllers.

In particular, removes the custom Response class for a more idiomatic approach, using nullability to signal NotFound and turns validation errors into
IllegalArgumentExceptions and then into BadRequest rather than an InternalServerError for less surprising error code.

* adds ktor plugin to build to simplify further, adjusts README to reflect different name of fatJar

* Adds the detekt plugin with the standard rules and fixes all warnings to make the code more familiar in layout and look.

In particular:
* moves code to directories corresponding to package
* avoids wildcard imports
* pulls constants out and uses `const` keyword for them
* fixes assorted formatting nags
* removes magic numbers

Additionally:
* Simplifies the `ElectricityReadingsGenerator`

* Fixes some table formatting niggles in README

* Restores explicit http status block with optional body in output from ktor.

This had been modelled as a pseudo domain class (`Response`) that wrapped header and body along with an extension function to the `ApplicationCall`. However, that
approach conflated http concerns and the chosen approach of wrapping up the domain concerns in a sort of JSON "envelope" with "normal" code.

As these concerns are all about how the domain logic is mapping into specific outputs, we now go for a simple Ktor plugin that takes responsibility for:
* mapping `null` responses to 404
* wrapping simple status code responses so that they can be serialised to JSON without the `ContentNegotiation` plugin intervening
* wrapping normal responses into a `statusCode` and `body`
* turning `IllegalArgumentExceptions` into 400 with an additional error message

`ApplicationTest` is extended to validate these cases.